Move print responsability to a new TransactionPrinter class
